client/requirements.txt
requests
websocket-client
python-socketio
addict
zenlog
